Use remark-shiki-twoslash and create html-to-jsx plugin to solve the issue
