+++ apps/docs/package.json
@@ -12,8 +12,8 @@
     "immer": "^10.0.2",
     "react": "18.2.0",
     "react-composable-form": "workspace:*",
-    "react-imperative-hook": "workspace:*",
     "react-dom": "18.2.0",
+    "react-imperative-hook": "workspace:*",
     "zod": "^3.21.4"
   },
   "devDependencies": {
@@ -29,7 +29,8 @@
     "@yas/env": "workspace:*",
     "@yas/tsconfig": "workspace:*",
     "docusaurus-plugin-sass": "^0.2.5",
-    "shiki-twoslash-remark": "workspace:*",
+    "remark-html-to-jsx": "workspace:*",
+    "remark-shiki-twoslash": "^3.1.3",
     "sass": "^1.66.1",
     "sass-loader": "^13.3.2",
     "shiki-twoslash-loader": "workspace:*",
